+++ resdk/resources/sample.py
@@ -49,8 +49,18 @@ class Sample(SampleUtilsMixin, BaseCollection):
 
     endpoint = 'sample'
 
+    WRITABLE_FIELDS = ('tags',) + BaseCollection.WRITABLE_FIELDS
+
     #: (lazy loaded) list of collections  to which object belongs
     _collections = None
+
+    def __init__(self, slug=None, id=None, model_data=None,  # pylint: disable=redefined-builtin
+                 resolwe=None):
+        """Initialize attributes."""
+        #: sample's tags
+        self.tags = None
+
+        super(Sample, self).__init__(slug, id, model_data, resolwe)
 
     def update(self):
         """Clear cache and update resource fields from the server."""